+++ test/win/CMakeLists.txt
@@ -1,13 +1,4 @@
 ##
-
-add_executable(pe_test
-  pe.cc
-)
-
-target_link_libraries(pe_test
-  belawin
-)
-
 add_executable(pecoff_test
   pecoff.cc
 )
